# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CSE 341 Project 2 API",
        "description": "REST API performing full CRUD operations on the contacts, users, colleges, courses, and instructors collections in MongoDB. All routes include data validation and error handling, and return proper HTTP status codes. Every POST, PUT, and DELETE route requires an authenticated GitHub session and returns 401 otherwise. To try a protected route in Swagger UI, open /login in a browser and sign in with the same host, then reload this page.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://web-servises-cse-341.onrender.com/",
            "description": "Production (Render)"
        },
        {
            "url": "http://localhost:3000",
            "description": "Local development"
        }
    ],
...
    "paths": {
...
        "/courses": {
            "get": {
                "tags": ["Courses"],
                "operationId": "getCourses",
                "summary": "Get all courses",
                "description": "Returns a list of every course in the courses collection.",
                "responses": {
                    "200": {
                        "description": "A list of courses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Course"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "An unexpected error occurred"
                    }
                }
            },
            "post": {
                "tags": ["Courses"],
                "operationId": "createCourse",
                "security": [{ "githubOAuth": [] }],
                "summary": "Create a course",
                "description": "Inserts a new course into the courses collection. collegeId must match an existing college.",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Course"
                            },
                            "examples": {
                                "course": {
                                    "value": {
                                        "name": "Bachelor of Science in Software Engineering",
                                        "code": "BSSE",
                                        "collegeId": "64f1e2d3a5b6c7d8e9f0a1b2",
                                        "credits": 130,
                                        "duration": 4,
                                        "level": "bachelors",
                                        "faculty": "Faculty of Technology",
                                        "description": "Software architecture, testing, and project management."
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Course created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "64f1e2d3a5b6c7d8e9f0a1b3"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Missing required fields, invalid data, or unknown collegeId"
                    },
                    "401": {
                        "description": "Not authenticated; log in with GitHub first"
                    },
                    "500": {
                        "description": "An unexpected error occurred"
                    }
                }
            }
        },
        "/courses/{id}": {
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "The id of the course",
                    "schema": {
                        "type": "string"
                    }
                }
            ],
            "get": {
                "tags": ["Courses"],
                "operationId": "getCourseById",
                "summary": "Get a course by id",
                "description": "Returns a single course matching the given id.",
                "responses": {
                    "200": {
                        "description": "The requested course",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Course"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid course id"
                    },
                    "404": {
                        "description": "Course not found"
                    },
                    "500": {
                        "description": "An unexpected error occurred"
                    }
                }
            },
            "put": {
                "tags": ["Courses"],
                "operationId": "updateCourse",
                "security": [{ "githubOAuth": [] }],
                "summary": "Update a course by id",
                "description": "Replaces the fields of the course matching the given id. All fields are required.",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Course"
                            },
                            "examples": {
                                "course": {
                                    "value": {
                                        "name": "Bachelor of Science in Software Engineering",
                                        "code": "BSSE",
                                        "collegeId": "64f1e2d3a5b6c7d8e9f0a1b2",
                                        "credits": 135,
                                        "duration": 4,
                                        "level": "bachelors",
                                        "faculty": "Faculty of Technology",
                                        "description": "Software architecture, testing, and project management."
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Course updated"
                    },
                    "400": {
                        "description": "Invalid course id, missing fields, or unknown collegeId"
                    },
                    "401": {
                        "description": "Not authenticated; log in with GitHub first"
                    },
                    "404": {
                        "description": "Course not found"
                    },
                    "500": {
                        "description": "An unexpected error occurred"
                    }
                }
            },
            "delete": {
                "tags": ["Courses"],
                "operationId": "deleteCourse",
                "security": [{ "githubOAuth": [] }],
                "summary": "Delete a course by id",
                "description": "Deletes the course matching the given id.",
                "responses": {
                    "204": {
                        "description": "Course deleted"
                    },
                    "400": {
                        "description": "Invalid course id"
                    },
                    "401": {
                        "description": "Not authenticated; log in with GitHub first"
                    },
                    "404": {
                        "description": "Course not found"
                    },
                    "500": {
                        "description": "An unexpected error occurred"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "githubOAuth": {
                "type": "oauth2",
                "flows": {
                    "authorizationCode": {
                        "authorizationUrl": "https://github.com/login/oauth/authorize",
                        "tokenUrl": "https://github.com/login/oauth/access_token",
                        "scopes": {
                            "user:email": "Read access to your email address"
                        }
                    }
                }
            }
        },
        "schemas": {
...
            "Course": {
                "type": "object",
                "required": ["name", "code", "collegeId", "credits", "duration", "level", "faculty", "description"],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Bachelor of Science in Computer Science"
                    },
                    "code": {
                        "type": "string",
                        "description": "Short unique code, 2-12 letters, numbers, or dashes.",
                        "example": "BSCS"
                    },
                    "collegeId": {
                        "type": "string",
                        "description": "The _id of the college that offers this course. Must match an existing college.",
                        "example": "64f1e2d3a5b6c7d8e9f0a1b2"
                    },
                    "credits": {
                        "type": "integer",
                        "description": "Total credit hours for the course.",
                        "example": 120
                    },
                    "duration": {
                        "type": "integer",
                        "description": "Length of the course in years.",
                        "example": 4
                    },
                    "level": {
                        "type": "string",
                        "enum": ["certificate", "diploma", "bachelors", "masters", "phd"],
                        "example": "bachelors"
                    },
                    "faculty": {
                        "type": "string",
                        "example": "Faculty of Technology"
                    },
                    "description": {
                        "type": "string",
                        "example": "Covers programming, algorithms, databases, and software engineering."
                    }
                }
            },
...
        }
    }
}